move tests and use event helper
